requirements.txt
# ---- Core ----
gitpython>=3.1.40
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0

# ---- Embeddings & LLM ----
openai>=1.12.0
sentence-transformers>=2.5.0

# ---- Vector Store ----
chromadb>=0.4.22

# ---- Graphs ----
networkx>=3.2

# ---- API ----
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# ---- UI ----
streamlit>=1.31.0
requests>=2.31.0

# ---- Dev / Utils ----
python-dotenv>=1.0.0
